@@ -4,38 +4,42 @@ import { UserController } from "../controllers/user/controller.user";
 import { QuestController } from "../controllers/quest/controller.quest";
 import { transformToGeoPostGIS } from "../utils/postGIS";
 import { error, output } from "../utils";
-import { publishQuestNotifications, QuestNotificationActions } from "../websocket/websocket.quest";
 import { QuestsResponseController } from "../controllers/quest/controller.questsResponse";
 import { MediaController } from "../controllers/controller.media";
 import { addUpdateReviewStatisticsJob } from "../jobs/updateReviewStatistics";
 import { updateQuestsStatisticJob } from "../jobs/updateQuestsStatistic";
-import { SkillsFiltersController } from "../controllers/controller.skillsFilters";
+import { QuestNotificationActions } from '../controllers/controller.broker';
 import {
+  DisputeStatus,
   Quest,
   QuestChat,
   QuestChatStatuses,
+  QuestDispute,
   QuestsResponse,
+  QuestsResponseStatus,
   QuestsResponseType,
   QuestStatus,
-  Review,
-  StarredQuests,
+  QuestsReview,
+  QuestsStarred,
   User,
   UserRole,
   QuestRaiseView,
   QuestRaiseStatus,
 } from "@workquest/database-models/lib/models";
 import { updateQuestRaiseViewStatusJob } from "../jobs/updateQuestRaiseViewStatus";
+import { SkillsFiltersController } from '../controllers/controller.skillsFilters';
 
-export const searchFields = [
-  "title",
-  "description",
+export const searchQuestFields = [
+  'title',
+  'description',
+  'locationPlaceName'
 ];
 
 export async function getQuest(r) {
   const user: User = r.auth.credentials;
 
   const include = [{
-    model: StarredQuests,
+    model: QuestsStarred,
     as: "star",
     where: { userId: r.auth.credentials.id },
     required: false
@@ -44,6 +48,28 @@ export async function getQuest(r) {
     as: "response",
     where: { workerId: r.auth.credentials.id },
     required: false
+  }, {
+    model: User.scope('shortWithWallet'),
+    as: 'user'
+  }, {
+    model: User.scope('shortWithWallet'),
+    as: 'assignedWorker'
+  }, {
+    model: QuestDispute.unscoped(),
+    as: 'openDispute',
+    where: {
+      [Op.or]: [
+        { opponentUserId: r.auth.credentials.id },
+        { openDisputeUserId: r.auth.credentials.id },
+      ],
+      status: { [Op.in]: [DisputeStatus.pending, DisputeStatus.inProgress] },
+    },
+    required: false,
+  }, {
+    model: QuestsReview.unscoped(),
+    as: 'yourReview',
+    where: { fromUserId: r.auth.credentials.id },
+    required: false,
   }] as any[];
 
   if (user.role === UserRole.Worker) {
@@ -51,8 +77,8 @@ export async function getQuest(r) {
       model: QuestChat.scope('idsOnly'),
       as: 'questChat',
       required: false,
-      where: { workerId: user.id }
-    })
+      where: { workerId: user.id },
+    });
   }
 
   const quest = await Quest.findOne({
@@ -61,7 +87,7 @@ export async function getQuest(r) {
   });
 
   if (!quest) {
-    return error(Errors.NotFound, "Quest not found", { questId: r.params.questId });
+    return error(Errors.NotFound, 'Quest not found', { questId: r.params.questId });
   }
 
   return output(quest);
@@ -80,18 +106,17 @@ export async function createQuest(r) {
   const quest = await Quest.create({
     userId: employer.id,
     status: QuestStatus.Created,
-    category: r.payload.category,
     workplace: r.payload.workplace,
     employment: r.payload.employment,
     priority: r.payload.priority,
-    location: r.payload.location,
     title: r.payload.title,
     description: r.payload.description,
     price: r.payload.price,
     medias: r.payload.medias,
     adType: r.payload.adType,
-    locationPlaceName: r.payload.locationPlaceName,
-    locationPostGIS: transformToGeoPostGIS(r.payload.location),
+    location: r.payload.locationFull.location,
+    locationPlaceName: r.payload.locationFull.locationPlaceName,
+    locationPostGIS: transformToGeoPostGIS(r.payload.locationFull.location),
   }, { transaction });
 
   const questController = new QuestController(quest);
@@ -107,9 +132,7 @@ export async function createQuest(r) {
     role: UserRole.Employer,
   });
 
-  return output(
-    await Quest.findByPk(quest.id)
-  )
+  return output(await Quest.findByPk(quest.id));
 }
 
 export async function editQuest(r) {
@@ -133,16 +156,27 @@ export async function editQuest(r) {
     title: r.payload.title,
     adType: r.payload.adType,
     priority: r.payload.priority,
-    category: r.payload.category,
     workplace: r.payload.workplace,
     employment: r.payload.employment,
     description: r.payload.description,
-    location: r.payload.location,
-    locationPlaceName: r.payload.locationPlaceName,
-    locationPostGIS: transformToGeoPostGIS(r.payload.location),
+    location: r.payload.locationFull.location,
+    locationPlaceName: r.payload.locationFull.locationPlaceName,
+    locationPostGIS: transformToGeoPostGIS(r.payload.locationFull.location),
   }, { transaction });
 
   await transaction.commit();
+
+  const responses = await QuestsResponse.findAll({
+    where: { questId: questController.quest.id, status: QuestsResponseStatus.Open },
+  });
+
+  if (responses.length !== 0) {
+    r.server.app.broker.sendQuestNotification({
+      action: QuestNotificationActions.questEdited,
+      recipients: responses.map(_ => _.workerId),
+      data: questController.quest,
+    });
+  }
 
   return output(questController.quest);
 }
@@ -218,10 +252,10 @@ export async function startQuest(r) {
 
   await transaction.commit();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: questController.quest,
     recipients: [assignedWorkerController.user.id],
-    action: QuestNotificationActions.questStarted,
+    action: QuestNotificationActions.waitWorker,
   });
 
   return output();
@@ -247,7 +281,7 @@ export async function rejectWorkOnQuest(r) {
 
   await transaction.commit();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     recipients: [questController.quest.userId],
     action: QuestNotificationActions.workerRejectedQuest,
     data: questController.quest,
@@ -281,7 +315,7 @@ export async function acceptWorkOnQuest(r) {
     role: UserRole.Worker,
   });
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: questController.quest,
     recipients: [questController.quest.userId],
     action: QuestNotificationActions.workerAcceptedQuest,
@@ -302,7 +336,7 @@ export async function completeWorkOnQuest(r) {
 
   await questController.completeWork();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: questController.quest,
     recipients: [questController.quest.userId],
     action: QuestNotificationActions.workerCompletedQuest,
@@ -323,7 +357,7 @@ export async function acceptCompletedWorkOnQuest(r) {
 
   await questController.approveCompletedWork();
 
-  await publishQuestNotifications(r.server, {
+  r.server.app.broker.sendQuestNotification({
     data: quest,
     recipients: [quest.assignedWorkerId],
     action: QuestNotificationActions.employerAcceptedCompletedQuest,
@@ -348,28 +382,7 @@ export async function acceptCompletedWorkOnQuest(r) {
   return output();
 }
 
-export async function rejectCompletedWorkOnQuest(r) {
-  const employer: User = r.auth.credentials;
-
-  const quest = await Quest.findByPk(r.params.questId);
-  const questController = new QuestController(quest);
-
-  questController
-    .employerMustBeQuestCreator(employer.id)
-    .questMustHaveStatus(QuestStatus.WaitConfirm)
-
-  await questController.rejectCompletedWork();
-
-  await publishQuestNotifications(r.server, {
-    data: quest,
-    recipients: [quest.assignedWorkerId],
-    action: QuestNotificationActions.employerRejectedCompletedQuest,
-  });
-
-  return output();
-}
-
-export async function getQuests(r) {
+export async function getPayloadQuests(r) {
   const user: User = r.auth.credentials;
 
   const entersAreaLiteral = literal(
@@ -388,9 +401,177 @@ export async function getQuests(r) {
     '(1 = (CASE WHEN EXISTS (SELECT * FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."path" IN (:path)) THEN 1 END))' +
     'OR (1 = (CASE WHEN EXISTS (SELECT * FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."industryKey" IN (:industryKey)) THEN 1 END))'
   );
-  const questRaiseViewLiteral = literal(
-    '(SELECT "type" FROM "QuestRaiseViews" WHERE "questId" = "Quest"."id" AND "QuestRaiseViews"."status" = 0)'
+  const userSearchLiteral = literal(
+    // TODO добавь эти поля в replace типо так ILIKE '%:searchByFirstName%'`
+    `(SELECT "firstName" FROM "Users" WHERE "id" = "Quest"."userId") ILIKE '%${r.payload.q}%'` +
+    `OR (SELECT "lastName" FROM "Users" WHERE "id" = "Quest"."userId") ILIKE '%${r.payload.q}%'`
   );
+  const questChatWorkerLiteral = literal(
+    '"questChat"."workerId" = "Quest"."assignedWorkerId"'
+  );
+
+  const order = [];
+  const include = [];
+  const replacements = {};
+  const where = {
+    [Op.and]: [],
+    ...(r.payload.adType && { adType: r.payload.adType }),
+    ...(r.payload.filter && { filter: r.params.filter }),
+    ...(r.params.userId && { userId: r.params.userId }),
+    ...(r.params.workerId && { assignedWorkerId: r.params.workerId }),
+    ...(r.payload.statuses && { status: { [Op.in]: r.payload.statuses } }),
+    ...(r.payload.performing && { assignedWorkerId: r.auth.credentials.id }),
+    ...(r.payload.priorities && { priority: { [Op.in]: r.payload.priorities } }),
+    ...(r.payload.workplaces && { workplace: { [Op.in]: r.payload.workplaces } }),
+    ...(r.payload.employments && { employment: { [Op.in]: r.payload.employments } }),
+    ...(r.payload.priceBetween && { price: { [Op.between]: [r.payload.priceBetween.from, r.payload.priceBetween.to] } }),
+  };
+
+  if (r.payload.q) {
+    where[Op.or] = searchQuestFields.map(field => ({
+      [field]: { [Op.iLike]: `%${r.payload.q}%` }
+    }));
+
+    where[Op.or].push(userSearchLiteral)
+  }
+  if (r.payload.specializations) { // TODO r.query.specialization on r.query.specialization[s]
+    const {
+      paths,
+      industryKeys
+    } = SkillsFiltersController.splitPathsAndSingleKeysOfIndustry(r.payload.specializations);
+
+    if (paths.length !== 0 && industryKeys.length === 0) {
+      replacements['path'] = paths;
+      where[Op.and].push(questSpecializationOnlyPathsLiteral);
+    }
+    if (paths.length === 0 && industryKeys.length !== 0) {
+      replacements['industryKey'] = industryKeys;
+      where[Op.and].push(questSpecializationOnlyIndustryKeysLiteral);
+    }
+    if (paths.length !== 0 && industryKeys.length !== 0) {
+      replacements['path'] = paths;
+      replacements['industryKey'] = industryKeys;
+      where[Op.and].push(questSpecializationIndustryKeysAndPathsLiteral);
+    }
+  }
+  if (r.payload.northAndSouthCoordinates) {
+    replacements['northLng'] = r.payload.northAndSouthCoordinates.north.longitude;
+    replacements['northLat'] = r.payload.northAndSouthCoordinates.north.latitude;
+    replacements['southLng'] = r.payload.northAndSouthCoordinates.south.longitude;
+    replacements['southLat'] = r.payload.northAndSouthCoordinates.south.latitude;
+
+    where[Op.and].push(entersAreaLiteral);
+  }
+  if (user.role === UserRole.Worker) {
+    include.push({
+      model: QuestChat.scope('idsOnly'),
+      where: { workerId: user.id },
+      as: 'questChat',
+      required: false,
+    });
+  }
+  if (user.role === UserRole.Employer) {
+    include.push({
+      model: QuestChat.scope('idsOnly'),
+      as: 'questChat',
+      attributes: {
+        include: [[questChatLiteral, 'id']],
+      },
+      where: { employerId: user.id, questChatWorkerLiteral },
+      required: false,
+      include: {
+        model: Quest.unscoped(),
+        as: 'quest',
+        attributes: ['id', 'status'],
+        where: {
+          status: [QuestStatus.Active, QuestStatus.Dispute, QuestStatus.WaitWorker, QuestStatus.WaitConfirm],
+        },
+        required: false,
+      },
+    });
+  }
+
+  include.push(
+    {
+      model: QuestsReview.unscoped(),
+      as: 'yourReview',
+      where: { fromUserId: r.auth.credentials.id },
+      required: false,
+    },
+    {
+      model: QuestsStarred.unscoped(),
+      as: 'star',
+      where: { userId: r.auth.credentials.id },
+      required: r.payload.starred,
+    },
+    {
+      model: QuestsResponse.unscoped(),
+      as: 'invited',
+      required: r.payload.invited,
+      where: {
+        [Op.and]: [{ workerId: r.auth.credentials.id }, { type: QuestsResponseType.Invite }],
+      },
+    },
+    {
+      model: QuestsResponse.unscoped(),
+      as: 'responded',
+      required: r.payload.responded,
+      where: {
+        [Op.and]: [{ workerId: r.auth.credentials.id }, { type: QuestsResponseType.Response }],
+      },
+    },
+  );
+
+  for (const [key, value] of Object.entries(r.payload.sort || {})) {
+    order.push([key, value]);
+  }
+
+  const { count, rows } = await Quest.findAndCountAll({
+    distinct: true,
+    limit: r.payload.limit,
+    offset: r.payload.offset,
+    include,
+    order,
+    where,
+    replacements,
+  });
+
+  return output({ count, quests: rows });
+}
+
+// TODO отрефракторить!
+export function getQuests(type: 'list' | 'points') {
+  return async function(r) {
+    const user: User = r.auth.credentials;
+
+    const entersAreaLiteral = literal(
+      'st_within("Quest"."locationPostGIS", st_makeenvelope(:northLng, :northLat, :southLng, :southLat, 4326))'
+    );
+    const questChatLiteral = literal(
+      'CASE WHEN "questChat->quest" = NULL THEN NULL ELSE "questChat->quest"."id" END'
+    );
+    const questSpecializationOnlyPathsLiteral = literal(
+      '(1 = (CASE WHEN EXISTS (SELECT "id" FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."path" IN (:path)) THEN 1 END))'
+    );
+    const questSpecializationOnlyIndustryKeysLiteral = literal(
+      '(1 = (CASE WHEN EXISTS (SELECT * FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."industryKey" IN (:industryKey)) THEN 1 END))'
+    );
+    const questSpecializationIndustryKeysAndPathsLiteral = literal(
+      '(1 = (CASE WHEN EXISTS (SELECT * FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."path" IN (:path)) THEN 1 END))' +
+      'OR (1 = (CASE WHEN EXISTS (SELECT * FROM "QuestSpecializationFilters" WHERE "questId" = "Quest"."id" AND "QuestSpecializationFilters"."industryKey" IN (:industryKey)) THEN 1 END))'
+    );
+    const userSearchLiteral = literal(
+      // TODO добавь эти поля в replace типо так ILIKE '%:searchByFirstName%'`
+      `(SELECT "firstName" FROM "Users" WHERE "id" = "Quest"."userId") ILIKE '%${r.query.q}%'` +
+      `OR (SELECT "lastName" FROM "Users" WHERE "id" = "Quest"."userId") ILIKE '%${r.query.q}%'`
+    );
+    const questChatWorkerLiteral = literal(
+      '"questChat"."workerId" = "Quest"."assignedWorkerId"'
+    );
+
+    const questRaiseViewLiteral = literal(
+      '(SELECT "type" FROM "QuestRaiseViews" WHERE "questId" = "Quest"."id" AND "QuestRaiseViews"."status" = 0)'
+    );
 
   const order = [[questRaiseViewLiteral, 'asc']] as any;
   const include = [];
@@ -409,122 +590,126 @@ export async function getQuests(r) {
     ...(r.query.priceBetween && { price: { [Op.between]: [r.query.priceBetween.from, r.query.priceBetween.to] } }),
   };
 
-  if (r.query.q) {
-    where[Op.or] = searchFields.map(field => ({
-      [field]: { [Op.iLike]: `%${r.query.q}%` }
-    }));
-  }
-  if (r.query.specializations) { // TODO r.query.specialization on r.query.specialization[s]
-    const { paths, industryKeys } = SkillsFiltersController.splitPathsAndSingleKeysOfIndustry(r.query.specializations);
+    if (r.query.q) {
+      where[Op.or] = searchQuestFields.map(field => ({
+        [field]: { [Op.iLike]: `%${r.query.q}%` }
+      }));
 
-    if (paths.length !== 0 && industryKeys.length === 0) {
-      replacements['path'] = paths;
-      where[Op.and].push(questSpecializationOnlyPathsLiteral);
+      where[Op.or].push(userSearchLiteral)
     }
-    if (paths.length === 0 && industryKeys.length !== 0) {
-      replacements['industryKey'] = industryKeys;
-      where[Op.and].push(questSpecializationOnlyIndustryKeysLiteral);
-    }
-    if (paths.length !== 0 && industryKeys.length !== 0) {
-      replacements['path'] = paths;
-      replacements['industryKey'] = industryKeys;
-      where[Op.and].push(questSpecializationIndustryKeysAndPathsLiteral);
-    }
-  }
-  if (r.query.north && r.query.south) {
-    replacements['northLng'] = r.query.north.longitude;
-    replacements['northLat'] = r.query.north.latitude;
-    replacements['southLng'] = r.query.south.longitude;
-    replacements['southLat'] = r.query.south.latitude;
+    if (r.query.specializations) { // TODO r.query.specialization on r.query.specialization[s]
+      const {
+        paths,
+        industryKeys
+      } = SkillsFiltersController.splitPathsAndSingleKeysOfIndustry(r.query.specializations);
 
-    where[Op.and].push(entersAreaLiteral);
-  }
-  if (user.role === UserRole.Worker) {
-    include.push({
-      model: QuestChat.scope('idsOnly'),
-      where: { workerId: user.id },
-      as: 'questChat',
-      required: false,
-    });
-  }
-  if (user.role === UserRole.Employer) {
-    include.push({
-      model: QuestChat.scope('idsOnly'),
-      as: 'questChat',
-      attributes: {
-        include: [[questChatLiteral, 'id']]
-      },
-      where: { employerId: user.id },
-      required: false,
-      include: {
-        model: Quest.unscoped(),
-        as: 'quest',
-        attributes: ["id", "status"],
-        where: {
-          status: [
-            QuestStatus.Active,
-            QuestStatus.Dispute,
-            QuestStatus.WaitWorker,
-            QuestStatus.WaitConfirm,
-          ],
+      if (paths.length !== 0 && industryKeys.length === 0) {
+        replacements['path'] = paths;
+        where[Op.and].push(questSpecializationOnlyPathsLiteral);
+      }
+      if (paths.length === 0 && industryKeys.length !== 0) {
+        replacements['industryKey'] = industryKeys;
+        where[Op.and].push(questSpecializationOnlyIndustryKeysLiteral);
+      }
+      if (paths.length !== 0 && industryKeys.length !== 0) {
+        replacements['path'] = paths;
+        replacements['industryKey'] = industryKeys;
+        where[Op.and].push(questSpecializationIndustryKeysAndPathsLiteral);
+      }
+    }
+    if (r.query.northAndSouthCoordinates) {
+      replacements['northLng'] = r.query.northAndSouthCoordinates.north.longitude;
+      replacements['northLat'] = r.query.northAndSouthCoordinates.north.latitude;
+      replacements['southLng'] = r.query.northAndSouthCoordinates.south.longitude;
+      replacements['southLat'] = r.query.northAndSouthCoordinates.south.latitude;
+
+      where[Op.and].push(entersAreaLiteral);
+    }
+    if (user.role === UserRole.Worker) {
+      include.push({
+        model: QuestChat.scope('idsOnly'),
+        where: { workerId: user.id },
+        as: 'questChat',
+        required: false,
+      });
+    }
+    if (user.role === UserRole.Employer) {
+      include.push({
+        model: QuestChat.scope('idsOnly'),
+        as: 'questChat',
+        attributes: {
+          include: [[questChatLiteral, 'id']],
         },
-        required: false
-      },
-    });
-  }
-
-  include.push({
-    model: Review.unscoped(),
-    as: "yourReview",
-    where: { fromUserId: r.auth.credentials.id },
-    required: false,
-  }, {
-    model: StarredQuests.unscoped(),
-    as: "star",
-    where: { userId: r.auth.credentials.id },
-    required: r.query.starred,
-  }, {
-    model: QuestsResponse.unscoped(),
-    as: 'invited',
-    required: r.query.invited,
-    where: {
-      [Op.and]: [
-        { workerId: r.auth.credentials.id },
-        { type: QuestsResponseType.Invite },
-      ]
+        where: { employerId: user.id, questChatWorkerLiteral },
+        required: false,
+        include: {
+          model: Quest.unscoped(),
+          as: 'quest',
+          attributes: ['id', 'status'],
+          where: {
+            status: [QuestStatus.Active, QuestStatus.Dispute, QuestStatus.WaitWorker, QuestStatus.WaitConfirm],
+          },
+          required: false,
+        },
+      });
     }
-  }, {
-    model: QuestsResponse.unscoped(),
-    as: "responded",
-    required: r.query.responded,
-    where: {
-      [Op.and]: [
-        { workerId: r.auth.credentials.id },
-        { type: QuestsResponseType.Response },
-      ]
-    },
-  }, {
-    model: QuestChat.scope('idsOnly'),
-    as: 'questChat',
-    required: false,
-  }, {
-    model: QuestRaiseView,
-    as: 'raiseView'
-  });
 
-  for (const [key, value] of Object.entries(r.query.sort)) {
-    order.push([key, value]);
+    include.push(
+      {
+        model: QuestsReview.unscoped(),
+        as: 'yourReview',
+        where: { fromUserId: r.auth.credentials.id },
+        required: false,
+      },
+      {
+        model: QuestsStarred.unscoped(),
+        as: 'star',
+        where: { userId: r.auth.credentials.id },
+        required: r.query.starred,
+      },
+      {
+        model: QuestsResponse.unscoped(),
+        as: 'invited',
+        required: r.query.invited,
+        where: {
+          [Op.and]: [{ workerId: r.auth.credentials.id }, { type: QuestsResponseType.Invite }],
+        },
+      },
+      {
+        model: QuestsResponse.unscoped(),
+        as: 'responded',
+        required: r.query.responded,
+        where: {
+          [Op.and]: [{ workerId: r.auth.credentials.id }, { type: QuestsResponseType.Response }],
+        },
+      },
+    );
+
+    for (const [key, value] of Object.entries(r.query.sort || {})) {
+      order.push([key, value]);
+    }
+
+    // TODO !!!!
+    if (type === 'list') {
+      const { count, rows } = await Quest.findAndCountAll({
+        distinct: true,
+        limit: r.query.limit,
+        offset: r.query.offset,
+        include,
+        order,
+        where,
+        replacements,
+      });
+
+      return output({ count, quests: rows });
+    } else if (type === 'points') {
+      const quests = await Quest.findAll({
+        include, order, where, replacements,
+      });
+
+      return output({ quests });
+    }
   }
-
-  const { count, rows } = await Quest.findAndCountAll({
-    distinct: true,
-    limit: r.query.limit,
-    offset: r.query.offset,
-    include, order, where,
-    replacements,
-  });
-
-  return output({ count, quests: rows });
 }
 
 export async function setStar(r) {
@@ -543,6 +728,40 @@ export async function removeStar(r) {
   await questController.removeStar(user);
 
   return output();
+}
+
+export async function getAvailableQuestsForWorker(r) {
+  const workerResponseLiteral = literal(
+    '1 = (CASE WHEN NOT EXISTS (SELECT "id" FROM "QuestsResponses" WHERE "QuestsResponses"."workerId"=:workerId ' +
+      'AND "QuestsResponses"."questId" = "Quest"."id") THEN 1 END)'
+  );
+
+  const worker = await User.findByPk(r.params.workerId);
+  const workerController = new UserController(worker);
+
+  const employer: User = r.auth.credentials;
+  const employerController = new UserController(employer);
+
+  employerController
+    .userMustHaveRole(UserRole.Employer)
+
+  workerController
+    .userMustHaveRole(UserRole.Worker)
+
+  const { count, rows } = await Quest.findAndCountAll({
+    distinct: true,
+    col: '"Quest"."id"',
+    where: {
+      userId: employer.id,
+      workerResponseLiteral,
+      status: QuestStatus.Created,
+    },
+    limit: r.query.limit,
+    offset: r.query.offset,
+    replacements: { workerId: worker.id }
+  });
+
+  return output({ count, quests: rows });
 }
 
 export async function activateRaiseView(r) {
