@@ -1,30 +1,29 @@
-import { Op } from "sequelize";
-import {error, output} from "../utils";
-import {Errors} from "../utils/errors";
-import {publishQuestNotifications, QuestNotificationActions} from "../websocket/websocket.quest";
-import { ChatNotificationActions, publishChatNotifications } from "../websocket/websocket.chat";
-import {QuestsResponseController} from "../controllers/quest/controller.questsResponse";
-import {QuestController} from "../controllers/quest/controller.quest";
-import {UserController}  from "../controllers/user/controller.user";
+import { Op } from 'sequelize';
+import { error, output } from '../utils';
+import { Errors } from '../utils/errors';
+import { ChatNotificationActions, QuestNotificationActions } from '../controllers/controller.broker';
+import { QuestsResponseController } from '../controllers/quest/controller.questsResponse';
+import { QuestController } from '../controllers/quest/controller.quest';
+import { UserController } from '../controllers/user/controller.user';
+import { MediaController } from '../controllers/controller.media';
 import {
-  User,
   Chat,
-  Quest,
-  Message,
-  ChatType,
-  UserRole,
-  QuestChat,
   ChatMember,
-  MessageType,
+  ChatType,
   InfoMessage,
-  QuestStatus,
+  Message,
   MessageAction,
-  QuestsResponse,
+  MessageType,
+  Quest,
+  QuestChat,
   QuestChatStatuses,
-  QuestsResponseType,
+  QuestsResponse,
   QuestsResponseStatus,
-} from "@workquest/database-models/lib/models";
-import { MediaController } from "../controllers/controller.media";
+  QuestsResponseType,
+  QuestStatus,
+  User,
+  UserRole,
+} from '@workquest/database-models/lib/models';
 
 export async function responseOnQuest(r) {
   let questResponse: QuestsResponse;
@@ -41,12 +40,12 @@ export async function responseOnQuest(r) {
     where: {
       workerId: worker.id,
       questId: questController.quest.id,
-      status: { [Op.ne]: QuestsResponseStatus.Accepted }
-    }
+      status: { [Op.ne]: QuestsResponseStatus.Accepted },
+    },
   });
 
   if (questResponse) {
-    return error(Errors.AlreadyAnswer, "You already answered quest", { questResponse });
+    return error(Errors.AlreadyAnswer, 'You already answered quest', { questResponse });
 
     // if (questResponse.previousStatus === QuestsResponseStatus.Rejected) {
     //   return error(Errors.Forbidden, "Client already rejected your response on quest", { questResponse });
@@ -60,13 +59,16 @@ export async function responseOnQuest(r) {
 
   const medias = await MediaController.getMedias(r.payload.medias);
 
-  questResponse = await QuestsResponse.create({
-    workerId: worker.id,
-    questId: quest.id,
-    message: r.payload.message,
-    status: QuestsResponseStatus.Open,
-    type: QuestsResponseType.Response,
-  }, { transaction });
+  questResponse = await QuestsResponse.create(
+    {
+      workerId: worker.id,
+      questId: quest.id,
+      message: r.payload.message,
+      status: QuestsResponseStatus.Open,
+      type: QuestsResponseType.Response,
+    },
+    { transaction },
+  );
 
   const questResponseController = new QuestsResponseController(questResponse);
 
@@ -78,7 +80,7 @@ export async function responseOnQuest(r) {
     senderUserId: worker.id,
     chatId: chat.id,
     type: MessageType.info,
-    number: 1, /** Because create */
+    number: 1 /** Because create */,
     createdAt: Date.now(),
   });
   const infoMessage = InfoMessage.build({
@@ -91,29 +93,32 @@ export async function responseOnQuest(r) {
     chatId: chat.id,
     text: r.payload.message,
     type: MessageType.message,
-    number: 2, /** Because create */
+    number: 2 /** Because create */,
     createdAt: Date.now() + 100,
   });
   const questChat = QuestChat.build({
-    employerId: quest.userId ,
+    employerId: quest.userId,
     workerId: worker.id,
     questId: quest.id,
     responseId: questResponse.id,
     chatId: chat.id,
   });
-  const members = ChatMember.bulkBuild([{
-    unreadCountMessages: 0, /** Because created */
-    chatId: chat.id,
-    userId: r.auth.credentials.id,
-    lastReadMessageId: firstInfoMessage.id, /** Because created,  */
-    lastReadMessageNumber: firstInfoMessage.number,
-  }, {
-    unreadCountMessages: 1, /** Because created */
-    chatId: chat.id,
-    userId: quest.userId,
-    lastReadMessageId: null, /** Because created */
-    lastReadMessageNumber: null,
-  }]);
+  const members = ChatMember.bulkBuild([
+    {
+      unreadCountMessages: 0 /** Because created */,
+      chatId: chat.id,
+      userId: r.auth.credentials.id,
+      lastReadMessageId: firstInfoMessage.id /** Because created,  */,
+      lastReadMessageNumber: firstInfoMessage.number,
+    },
+    {
+      unreadCountMessages: 1 /** Because created */,
+      chatId: chat.id,
+      userId: quest.userId,
+      lastReadMessageId: null /** Because created */,
+      lastReadMessageNumber: null,
+    },
+  ]);
 
   chat.lastMessageId = responseWorkerMessage.id;
   chat.lastMessageDate = responseWorkerMessage.createdAt;
@@ -124,24 +129,32 @@ export async function responseOnQuest(r) {
     infoMessage.save({ transaction }),
     responseWorkerMessage.save({ transaction }),
     questChat.save({ transaction }),
-    ...members.map(member => member.save({ transaction })),
+    ...members.map((member) => member.save({ transaction })),
   ] as Promise<any>[]);
 
   await responseWorkerMessage.$set('medias', medias, { transaction });
 
   await transaction.commit();
 
-  await publishChatNotifications(r.server, {
+  r.server.app.broker.sendChatNotification({
     action: ChatNotificationActions.newMessage,
     recipients: [quest.userId],
     data: await Message.findByPk(firstInfoMessage.id),
+  });
+
+  questResponse.setDataValue('quest', quest);
+  questResponse.setDataValue('worker', workerController.shortCredentials);
+  r.server.app.broker.sendQuestNotification({
+    action: QuestNotificationActions.workerRespondedToQuest,
+    recipients: [quest.userId],
+    data: questResponse,
   });
 
   return output(chat);
 }
 
 export async function inviteOnQuest(r) {
-  let questResponse: QuestsResponse
+  let questResponse: QuestsResponse;
   const employer: User = r.auth.credentials;
   const employerController = new UserController(employer);
 
@@ -161,11 +174,12 @@ export async function inviteOnQuest(r) {
     where: {
       questId: questController.quest.id,
       workerId: invitedWorkerController.user.id,
-      status: { [Op.ne]: QuestsResponseStatus.Accepted} },
+      status: { [Op.ne]: QuestsResponseStatus.Accepted },
+    },
   });
 
   if (questResponse) {
-    return error(Errors.AlreadyAnswer, "You have already been invited user to the quest", { questResponse });
+    return error(Errors.AlreadyAnswer, 'You have already been invited user to the quest', { questResponse });
     // if(questResponse.previousStatus === QuestsResponseStatus.Rejected) {
     //   return error(Errors.Forbidden, 'Person reject quest invitation', {});
     // }
@@ -176,21 +190,24 @@ export async function inviteOnQuest(r) {
 
   const transaction = await r.server.app.db.transaction();
 
-  questResponse = await QuestsResponse.create({
-    workerId: invitedWorkerController.user.id,
-    questId: questController.quest.id,
-    message: r.payload.message,
-    status: QuestsResponseStatus.Open,
-    previousStatus: QuestsResponseStatus.Open,
-    type: QuestsResponseType.Invite,
-  }, { transaction });
+  questResponse = await QuestsResponse.create(
+    {
+      workerId: invitedWorkerController.user.id,
+      questId: questController.quest.id,
+      message: r.payload.message,
+      status: QuestsResponseStatus.Open,
+      previousStatus: QuestsResponseStatus.Open,
+      type: QuestsResponseType.Invite,
+    },
+    { transaction },
+  );
 
   const chat = Chat.build({ type: ChatType.quest });
   const firstInfoMessage = Message.build({
     senderUserId: employer.id,
     chatId: chat.id,
     type: MessageType.info,
-    number: 1, /** Because create */
+    number: 1 /** Because create */,
     createdAt: Date.now(),
   });
   const infoMessage = InfoMessage.build({
@@ -203,24 +220,27 @@ export async function inviteOnQuest(r) {
     chatId: chat.id,
     text: r.payload.message,
     type: MessageType.message,
-    number: 2, /** Because create */
+    number: 2 /** Because create */,
     createdAt: Date.now() + 100,
   });
-  const members = ChatMember.bulkBuild([{
-    unreadCountMessages: 0, /** Because created */
-    chatId: chat.id,
-    userId: r.auth.credentials.id,
-    lastReadMessageId: firstInfoMessage.id, /** Because created,  */
-    lastReadMessageNumber: 1
-  }, {
-    unreadCountMessages: 1, /** Because created */
-    chatId: chat.id,
-    userId: invitedWorker.id,
-    lastReadMessageId: null, /** Because created */
-    lastReadMessageNumber: null,
-  }]);
+  const members = ChatMember.bulkBuild([
+    {
+      unreadCountMessages: 0 /** Because created */,
+      chatId: chat.id,
+      userId: r.auth.credentials.id,
+      lastReadMessageId: firstInfoMessage.id /** Because created,  */,
+      lastReadMessageNumber: 1,
+    },
+    {
+      unreadCountMessages: 1 /** Because created */,
+      chatId: chat.id,
+      userId: invitedWorker.id,
+      lastReadMessageId: null /** Because created */,
+      lastReadMessageNumber: null,
+    },
+  ]);
   const questChat = QuestChat.build({
-    employerId: quest.userId ,
+    employerId: quest.userId,
     workerId: invitedWorker.id,
     questId: quest.id,
     responseId: questResponse.id,
@@ -236,15 +256,24 @@ export async function inviteOnQuest(r) {
     infoMessage.save({ transaction }),
     inviteEmployerMessage.save({ transaction }),
     questChat.save({ transaction }),
-    ...members.map( member => member.save({ transaction }) ),
+    ...members.map((member) => member.save({ transaction })),
   ] as Promise<any>[]);
 
   await transaction.commit();
 
-  await publishChatNotifications(r.server, {
+  questResponse.setDataValue('quest', questController.quest);
+  questResponse.setDataValue('employer', employerController.shortCredentials);
+
+  r.server.app.broker.sendChatNotification({
     action: ChatNotificationActions.newMessage,
     recipients: [quest.userId],
     data: await Message.findByPk(firstInfoMessage.id),
+  });
+
+  r.server.app.broker.sendQuestNotification({
+    action: QuestNotificationActions.employerInvitedWorkerToQuest,
+    recipients: [invitedWorker.id],
+    data: questResponse,
   });
 
   return output({ chat });
@@ -259,11 +288,17 @@ export async function userResponsesToQuest(r) {
   await questController.employerMustBeQuestCreator(employer.id);
 
   const { rows, count } = await QuestsResponse.findAndCountAll({
-    include: {
-      model: QuestChat.unscoped(),
-      attributes: ["chatId"],
-      as: 'questChat'
-    },
+    include: [
+      {
+        model: QuestChat.unscoped(),
+        attributes: ['chatId'],
+        as: 'questChat',
+      },
+      {
+        model: User.scope('shortWithWallet'),
+        as: 'worker',
+      },
+    ],
     where: { questId: questController.quest.id },
     limit: r.query.limit,
     offset: r.query.offset,
@@ -280,14 +315,17 @@ export async function responsesToQuestsForUser(r) {
 
   const { rows, count } = await QuestsResponse.findAndCountAll({
     where: { workerId: worker.id },
-    include: [{
-      model: Quest,
-      as: 'quest',
-    }, {
-      model: QuestChat.unscoped(),
-      attributes: ["chatId"],
-      as: 'questChat'
-    }],
+    include: [
+      {
+        model: Quest,
+        as: 'quest',
+      },
+      {
+        model: QuestChat.unscoped(),
+        attributes: ['chatId'],
+        as: 'questChat',
+      },
+    ],
     limit: r.query.limit,
     offset: r.query.offset,
   });
@@ -308,20 +346,24 @@ export async function acceptInviteOnQuest(r) {
   questsResponseController
     .workerMustBeInvitedToQuest(worker.id)
     .questsResponseMustHaveType(QuestsResponseType.Invite)
-    .questsResponseMustHaveStatus(QuestsResponseStatus.Open)
+    .questsResponseMustHaveStatus(QuestsResponseStatus.Open);
 
   const transaction = await r.server.app.db.transaction();
 
-  questResponse = await questResponse.update({
-    status: QuestsResponseStatus.Accepted,
-  }, { transaction });
+  questResponse = await questResponse.update(
+    {
+      status: QuestsResponseStatus.Accepted,
+    },
+    { transaction },
+  );
   questsResponseController.questsResponse = questResponse;
 
   const questChat = await QuestChat.findOne({
     where: { responseId: questResponse.id },
     include: {
-      model: Chat.unscoped(), as: 'chat',
-      include: [{ model: Message.unscoped(), as: 'lastMessage' }]
+      model: Chat.unscoped(),
+      as: 'chat',
+      include: [{ model: Message.unscoped(), as: 'lastMessage' }],
     },
   });
   const message = Message.build({
@@ -345,10 +387,10 @@ export async function acceptInviteOnQuest(r) {
 
   await transaction.commit();
 
-  await publishQuestNotifications(r.server, {
-    data: questResponse,
-    recipients: [questResponse.quest.userId],
+  r.server.app.broker.sendQuestNotification({
     action: QuestNotificationActions.workerAcceptedInvitationToQuest,
+    recipients: [questResponse.quest.userId],
+    data: questResponse,
   });
 
   return output();
@@ -367,24 +409,30 @@ export async function rejectInviteOnQuest(r) {
   questsResponseController
     .workerMustBeInvitedToQuest(worker.id)
     .questsResponseMustHaveType(QuestsResponseType.Invite)
-    .questsResponseMustHaveStatus(QuestsResponseStatus.Open)
+    .questsResponseMustHaveStatus(QuestsResponseStatus.Open);
 
   const transaction = await r.server.app.db.transaction();
 
-  questResponse = await questResponse.update({
-    status: QuestsResponseStatus.Rejected
-  }, { transaction });
+  questResponse = await questResponse.update(
+    {
+      status: QuestsResponseStatus.Rejected,
+    },
+    { transaction },
+  );
   questsResponseController.questsResponse = questResponse;
 
   //TODO In chat controller
   const questChat = await QuestChat.findOne({
     where: { responseId: questResponse.id },
     include: {
-      model: Chat.unscoped(), as: 'chat',
-      include: [{
-        model: Message.unscoped(),
-        as: 'lastMessage'
-      }]
+      model: Chat.unscoped(),
+      as: 'chat',
+      include: [
+        {
+          model: Message.unscoped(),
+          as: 'lastMessage',
+        },
+      ],
     },
   });
   const message = Message.build({
@@ -408,10 +456,10 @@ export async function rejectInviteOnQuest(r) {
 
   await transaction.commit();
 
-  await publishQuestNotifications(r.server, {
-    data: questResponse,
-    recipients: [questResponse.quest.userId],
+  r.server.app.broker.sendQuestNotification({
     action: QuestNotificationActions.workerRejectedInvitationToQuest,
+    recipients: [questResponse.quest.userId],
+    data: questResponse,
   });
 
   return output();
@@ -428,25 +476,29 @@ export async function rejectResponseOnQuest(r) {
 
   await questController.employerMustBeQuestCreator(employer.id);
 
-  questsResponseController
-    .questsResponseMustHaveType(QuestsResponseType.Response)
-    .questsResponseMustHaveStatus(QuestsResponseStatus.Open)
+  questsResponseController.questsResponseMustHaveType(QuestsResponseType.Response).questsResponseMustHaveStatus(QuestsResponseStatus.Open);
 
   const transaction = await r.server.app.db.transaction();
 
-  questResponse = await questResponse.update({
-    status: QuestsResponseStatus.Rejected
-  }, { transaction });
+  questResponse = await questResponse.update(
+    {
+      status: QuestsResponseStatus.Rejected,
+    },
+    { transaction },
+  );
   questsResponseController.questsResponse = questResponse;
 
   const questChat = await QuestChat.findOne({
     where: { responseId: questResponse.id },
     include: {
-      model: Chat.unscoped(), as: 'chat',
-      include: [{
-        model: Message.unscoped(),
-        as: 'lastMessage'
-      }]
+      model: Chat.unscoped(),
+      as: 'chat',
+      include: [
+        {
+          model: Message.unscoped(),
+          as: 'lastMessage',
+        },
+      ],
     },
   });
   const message = Message.build({
@@ -470,10 +522,10 @@ export async function rejectResponseOnQuest(r) {
 
   await transaction.commit();
 
-  await publishQuestNotifications(r.server, {
-    data: questResponse,
-    recipients: [questResponse.quest.userId],
+  r.server.app.broker.sendQuestNotification({
     action: QuestNotificationActions.employerRejectedWorkersResponse,
+    recipients: [questResponse.workerId],
+    data: questResponse,
   });
 
   return output();
