@@ -1,13 +1,13 @@
-import {error} from "../../utils";
-import {Transaction} from "sequelize";
-import {Errors} from "../../utils/errors";
-import {SkillsFiltersController} from "../controller.skillsFilters";
+import { error } from '../../utils';
+import { Transaction } from 'sequelize';
+import { Errors } from '../../utils/errors';
+import { SkillsFiltersController } from "../controller.skillsFilters";
 import {
   User,
   Quest,
   QuestChat,
   QuestStatus,
-  StarredQuests,
+  QuestsStarred,
   QuestRaiseView,
   QuestsResponse,
   QuestRaiseStatus,
@@ -18,14 +18,14 @@ abstract class QuestHelper {
   public abstract quest: Quest;
 
   public setStar(user: User) {
-    return StarredQuests.findOrCreate({
+    return QuestsStarred.findOrCreate({
       where: { userId: user.id, questId: this.quest.id },
       defaults: { userId: user.id, questId: this.quest.id },
     });
   }
 
   public removeStar(user: User) {
-    return StarredQuests.destroy({
+    return QuestsStarred.destroy({
       where: { userId: user.id, questId: this.quest.id },
     });
   }
@@ -64,9 +64,9 @@ abstract class QuestHelper {
   }
 
   /** Checks list */
-  public employerMustBeQuestCreator(userId: String): QuestHelper {
+  public employerMustBeQuestCreator(userId: string): QuestHelper {
     if (this.quest.userId !== userId) {
-      throw error(Errors.Forbidden, "User is not quest creator", {
+      throw error(Errors.Forbidden, 'User is not quest creator', {
         current: this.quest.userId,
         mustHave: userId,
       });
@@ -88,7 +88,7 @@ abstract class QuestHelper {
 
   public workerMustBeAppointedOnQuest(workerId: string): QuestHelper {
     if (this.quest.assignedWorkerId !== workerId) {
-      throw error(Errors.Forbidden, "Worker is not appointed on quest", {
+      throw error(Errors.Forbidden, 'Worker is not appointed on quest', {
         current: this.quest.userId,
         mustHave: workerId,
       });
@@ -99,7 +99,7 @@ abstract class QuestHelper {
 
   public userMustBelongToQuest(userId: string): QuestHelper {
     if (userId !== this.quest.userId && userId !== this.quest.assignedWorkerId) {
-      throw error(Errors.Forbidden, "User does not belong to quest", {});
+      throw error(Errors.Forbidden, 'User does not belong to quest', {});
     }
 
     return this;
@@ -107,22 +107,23 @@ abstract class QuestHelper {
 }
 
 export class QuestController extends QuestHelper {
-  constructor(
-    public quest: Quest,
-  ) {
+  constructor(public quest: Quest) {
     super();
 
     if (!quest) {
-      throw error(Errors.NotFound, "Quest not found", {});
+      throw error(Errors.NotFound, 'Quest not found', {});
     }
   }
 
   public async start(assignedWorker: User, transaction?: Transaction) {
     try {
-      this.quest = await this.quest.update({
-        assignedWorkerId: assignedWorker.id,
-        status: QuestStatus.WaitWorker,
-      }, { transaction });
+      this.quest = await this.quest.update(
+        {
+          assignedWorkerId: assignedWorker.id,
+          status: QuestStatus.WaitWorker,
+        },
+        { transaction },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -133,9 +134,12 @@ export class QuestController extends QuestHelper {
 
   public async close(transaction?: Transaction) {
     try {
-      this.quest = await this.quest.update({
-        status: QuestStatus.Closed,
-      }, { transaction });
+      this.quest = await this.quest.update(
+        {
+          status: QuestStatus.Closed,
+        },
+        { transaction },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -146,9 +150,12 @@ export class QuestController extends QuestHelper {
 
   public async completeWork(transaction?: Transaction) {
     try {
-      this.quest = await this.quest.update({
-        status: QuestStatus.WaitConfirm,
-      }, { transaction });
+      this.quest = await this.quest.update(
+        {
+          status: QuestStatus.WaitConfirm,
+        },
+        { transaction },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -159,9 +166,12 @@ export class QuestController extends QuestHelper {
 
   public async approveCompletedWork(transaction?: Transaction) {
     try {
-      this.quest = await this.quest.update({
-        status: QuestStatus.Done,
-      }, { transaction });
+      this.quest = await this.quest.update(
+        {
+          status: QuestStatus.Done,
+        },
+        { transaction },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -172,9 +182,12 @@ export class QuestController extends QuestHelper {
 
   public async rejectCompletedWork(transaction?: Transaction) {
     try {
-      this.quest = await this.quest.update({
-        status: QuestStatus.Dispute,
-      }, { transaction });
+      this.quest = await this.quest.update(
+        {
+          status: QuestStatus.Dispute,
+        },
+        { transaction },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -186,14 +199,21 @@ export class QuestController extends QuestHelper {
   async answerWorkOnQuest(worker: User, acceptWork: boolean, transaction: Transaction) {
     try {
       if (acceptWork) {
-        this.quest = await this.quest.update({
-          status: QuestStatus.Active,
-        }, { transaction });
+        this.quest = await this.quest.update(
+          {
+            status: QuestStatus.Active,
+          },
+          { transaction },
+        );
+        await Quest.update({ startedAt: Date.now() }, { where: { id: this.quest.id }, transaction });
       } else {
-        this.quest = await this.quest.update({
-          status: QuestStatus.Created,
-          assignedWorkerId: null,
-        }, { transaction });
+        this.quest = await this.quest.update(
+          {
+            status: QuestStatus.Created,
+            assignedWorkerId: null,
+          },
+          { transaction },
+        );
       }
     } catch (e) {
       if (transaction) {
@@ -207,7 +227,7 @@ export class QuestController extends QuestHelper {
     try {
       await QuestChat.destroy({ where: { questId: this.quest.id }, transaction });
       await QuestSpecializationFilter.destroy({ where: { questId: this.quest.id }, transaction });
-      await QuestsResponse.destroy({ where: { questId: this.quest.id }, transaction })
+      await QuestsResponse.destroy({ where: { questId: this.quest.id }, transaction });
       await this.quest.destroy({ force: true, transaction });
     } catch (e) {
       if (transaction) {
@@ -227,6 +247,8 @@ export class QuestController extends QuestHelper {
       throw e;
     }
   }
+
+  public static get(id) {}
 
   public async createRaiseView(userId: string, transaction: Transaction) {
     await QuestRaiseView.create({
