@@ -1,25 +1,31 @@
-import { addJob } from "../utils/scheduler";
-import { ChatMember, ChatsStatistic } from "@workquest/database-models/lib/models";
-import { Op } from "sequelize";
+import { addJob } from '../utils/scheduler';
+import { ChatMember, ChatsStatistic } from '@workquest/database-models/lib/models';
+import { Op } from 'sequelize';
 
 export type UserUnreadChatsPayload = {
-  userIds: string[],
-}
+  userIds: string[];
+};
 
 export async function updateCountUnreadChatsJob(payload: UserUnreadChatsPayload) {
-  return addJob("updateCountUnreadChats", payload);
+  return addJob('updateCountUnreadChats', payload);
 }
 
 export default async function updateCountUnreadChats(payload: UserUnreadChatsPayload) {
-  for (const id of payload.userIds) {
+  for (const userId of payload.userIds) {
     const unreadChatsCounter = await ChatMember.unscoped().count({
       where: {
-        userId: id,
+        userId,
         unreadCountMessages: { [Op.ne]: 0 },
       },
     });
 
-    await ChatsStatistic.update({ unreadCountChats: unreadChatsCounter }, { where: { userId: id } });
+    const [chatsStatistic, isCreated] = await ChatsStatistic.findOrCreate({
+      where: { userId },
+      defaults: { userId, unreadCountChats: unreadChatsCounter },
+    });
+
+    if (!isCreated) {
+      await chatsStatistic.update({ unreadCountChats: unreadChatsCounter });
+    }
   }
 }
-
